Add pParams to env
