+++ src/Server/Internal.hs
@@ -13,22 +13,19 @@ module Server.Internal
     , runAppM
     , getQueueRef
     , loadEnv
-    , checkForCleanUtxos
     ) where
 
-import           Control.Monad          (void, when)
 import           Control.Monad.Catch    (MonadThrow, MonadCatch)
 import           Control.Monad.IO.Class (MonadIO)
 import           Control.Monad.Reader   (ReaderT(ReaderT, runReaderT), MonadReader, asks)
+import           Data.Default           (def)
 import           Data.IORef             (newIORef)
 import           Data.Sequence          (empty)
-import           IO.ChainIndex          (getWalletUtxos)
-import           IO.Wallet              (HasWallet(..), getWalletAddr)
+import           IO.Wallet              (HasWallet(..))
+import           Ledger                 (Params(..), pParamsFromProtocolParams)
 import           Servant                (Handler)
 import           Server.Class           (Env(..), Queue, QueueRef, QueueElem, HasServer(..))
 import           Server.Config          (Config(..), configFile, decodeOrErrorFromFile)
-import           Server.Tx              (mkWalletTxOutRefs)
-import           Utils.ChainIndex       (filterCleanUtxos)
 import           Utils.Logger           (HasLogger(..))
 
 newtype NetworkM s a = NetworkM { unNetworkM :: ReaderT (Env s) Handler a }
@@ -51,11 +48,13 @@ getQueueRef = asks envQueueRef
 
 loadEnv :: forall s. HasServer s => IO (Env s)
 loadEnv = do
-    Config{..} <- decodeOrErrorFromFile configFile
-    let envMinUtxosAmount = cMinUtxosAmount
+    Config{..}   <- decodeOrErrorFromFile configFile
     envQueueRef  <- newIORef empty
     envWallet    <- decodeOrErrorFromFile cWalletFile
     envAuxiliary <- loadAuxiliaryEnv @s cAuxiliaryEnvFile
+    pp           <- decodeOrErrorFromFile "testnet/protocol-parameters.json"
+    let envMinUtxosAmount = cMinUtxosAmount
+        envLedgerParams   = Params def (pParamsFromProtocolParams pp) cNetworkId
     pure Env{..}
 
 newtype AppM s a = AppM { unAppM :: ReaderT (Env s) IO a }
@@ -69,12 +68,3 @@ instance HasLogger (AppM s) where
 
 instance HasWallet (AppM s) where
     getRestoredWallet = asks envWallet
-
-checkForCleanUtxos :: (HasWallet m, HasLogger m, MonadReader (Env s) m) => m ()
-checkForCleanUtxos = do
-    addr       <- getWalletAddr
-    cleanUtxos <- length . filterCleanUtxos <$> getWalletUtxos
-    minUtxos   <- asks envMinUtxosAmount
-    when (cleanUtxos < minUtxos) $ do
-        logMsg "Address doesn't has enough clean UTXO's."
-        void $ mkWalletTxOutRefs addr (cleanUtxos - minUtxos)
